+++ src/components/ui/Loader/Loader.jsx
@@ -1,0 +1,11 @@
+import styles from "./Loader.module.scss";
+export const Loader = () => {
+  return (
+    <div className={styles["loader-container"]}>
+      <div className={styles["lds-ripple"]}>
+        <div></div>
+        <div></div>
+      </div>
+    </div>
+  );
+};
